FEAT: adding scripts et methods for cron Job
