+++ package.json
@@ -19,12 +19,14 @@
     "dev:strategyService:backend": "NODE_ENV=dev ts-node-dev --ignore-watch node_modules  ./src/apps/strategyService/start.ts",
     "dev:mechoui3:backend": "NODE_ENV=dev ts-node-dev --ignore-watch node_modules  ./src/apps/mechoui3/start.ts",
     "dev:backoffice:backend": "NODE_ENV=dev ts-node-dev --ignore-watch node_modules ./src/apps/backoffice/start.ts",
+    "dev:pricingService:backend": "NODE_ENV=dev ts-node-dev --ignore-watch node_modules ./src/apps/pricingService/start.ts",
     "dev:example:backend": "NODE_ENV=dev ts-node-dev --ignore-watch node_modules ./src/apps/example/start.ts",
     "windows:dev:mooc:backend": "set NODE_ENV=dev & ts-node-dev --ignore-watch node_modules  ./src/apps/mooc/start.ts",
     "windows:dev:mechoui:backend": "set NODE_ENV=dev & ts-node-dev --ignore-watch node_modules  ./src/apps/alertService/start.ts",
     "windows:dev:strategyService:backend": "set NODE_ENV=dev & ts-node-dev --ignore-watch node_modules  ./src/apps/strategyService/start.ts",
     "windows:dev:mechoui3:backend": "set NODE_ENV=dev & ts-node-dev --ignore-watch node_modules  ./src/apps/mechoui3/start.ts",
     "windows:dev:backoffice:backend": "set NODE_ENV=dev & ts-node-dev --ignore-watch node_modules ./src/apps/backoffice/start.ts",
+    "windows:dev:pricingservice:backend": "set NODE_ENV=dev & ts-node-dev --ignore-watch node_modules ./src/apps/pricingService/start.ts",
     "windows:dev:example:backend": "set NODE_ENV=dev & ts-node-dev --ignore-watch node_modules ./src/apps/example/start.ts",
     "lint": "prettier --write src/**/*.ts{,x}",
     "test": "npm run test:unit && npm run test:features",
