--- src/Contexts/mechoui3/Courses/application/SearchAll/SearchAllCoursesQuery.ts
@@ -1,3 +1,0 @@
-import { Query } from '../../../../Shared/domain/Query';
-
-export class SearchAllCoursesQuery implements Query {}
