--- src/Contexts/mechoui3/Courses/domain/BackofficeCourseName.ts
@@ -1,3 +1,0 @@
-import { StringValueObject } from '../../../Shared/domain/value-object/StringValueObject';
-
-export class BackofficeCourseName extends StringValueObject {}
