+++ src/Contexts/pricingService/application/Create/Boom3.ts
@@ -1,6 +1,6 @@
-import { CourseCreatedDomainEvent } from '../../../../Mooc/Courses/domain/CourseCreatedDomainEvent';
-import { DomainEventClass } from '../../../../Shared/domain/DomainEvent';
-import { DomainEventSubscriber } from '../../../../Shared/domain/DomainEventSubscriber';
+import { CourseCreatedDomainEvent } from '../../../Mooc/Courses/domain/CourseCreatedDomainEvent';
+import { DomainEventClass } from '../../../Shared/domain/DomainEvent';
+import { DomainEventSubscriber } from '../../../Shared/domain/DomainEventSubscriber';
 
 export class Boom3 implements DomainEventSubscriber<CourseCreatedDomainEvent> {
   subscribedTo(): DomainEventClass[] {
