+++ src/Contexts/pricingService/application/Create/CreateBackofficeCourseOnCourseCreated3.ts
@@ -1,6 +1,6 @@
-import { CourseCreatedDomainEvent } from '../../../../Mooc/Courses/domain/CourseCreatedDomainEvent';
-import { DomainEventClass } from '../../../../Shared/domain/DomainEvent';
-import { DomainEventSubscriber } from '../../../../Shared/domain/DomainEventSubscriber';
+import { CourseCreatedDomainEvent } from '../../../Mooc/Courses/domain/CourseCreatedDomainEvent';
+import { DomainEventClass } from '../../../Shared/domain/DomainEvent';
+import { DomainEventSubscriber } from '../../../Shared/domain/DomainEventSubscriber';
 //import { BackofficeCourseCreator } from './BackofficeCourseCreator';
 
 export class CreateBackofficeCourseOnCourseCreated3 implements DomainEventSubscriber<CourseCreatedDomainEvent> {
