+++ src/Contexts/pricingService/application/SearchAll/CoursesFinder.ts
@@ -1,7 +1,7 @@
 import { BackofficeCourseRepository } from '../../domain/BackofficeCourseRepository';
 
 export class CoursesFinder {
-  constructor(private coursesRepository: BackofficeCourseRepository) {}
+  constructor(private coursesRepository: BackofficeCourseRepository) { }
 
   async run() {
     const courses = await this.coursesRepository.searchAll();
