+++ src/Contexts/pricingService/application/SearchAll/SearchAllCoursesQuery.ts
@@ -1,0 +1,3 @@
+import { Query } from '../../../Shared/domain/Query';
+
+export class SearchAllCoursesQuery implements Query { }
