+++ src/Contexts/pricingService/application/SearchByCriteria/CoursesByCriteriaSearcher.ts
@@ -1,11 +1,11 @@
-import { Criteria } from '../../../../Shared/domain/criteria/Criteria';
-import { Filters } from '../../../../Shared/domain/criteria/Filters';
-import { Order } from '../../../../Shared/domain/criteria/Order';
+import { Criteria } from '../../../Shared/domain/criteria/Criteria';
+import { Filters } from '../../../Shared/domain/criteria/Filters';
+import { Order } from '../../../Shared/domain/criteria/Order';
 import { BackofficeCourseRepository } from '../../domain/BackofficeCourseRepository';
 import { BackofficeCoursesResponse } from '../BackofficeCoursesResponse';
 
 export class CoursesByCriteriaSearcher {
-  constructor(private repository: BackofficeCourseRepository) {}
+  constructor(private repository: BackofficeCourseRepository) { }
 
   async run(filters: Filters, order: Order, limit?: number, offset?: number): Promise<BackofficeCoursesResponse> {
     const criteria = new Criteria(filters, order, limit, offset);
