+++ src/Contexts/pricingService/application/SearchByCriteria/SearchCoursesByCriteriaQuery.ts
@@ -1,4 +1,4 @@
-import { Query } from '../../../../Shared/domain/Query';
+import { Query } from '../../../Shared/domain/Query';
 
 export class SearchCoursesByCriteriaQuery implements Query {
   readonly filters: Array<Map<string, string>>;
