+++ src/Contexts/pricingService/application/SearchByCriteria/SearchCoursesByCriteriaQueryHandler.ts
@@ -1,15 +1,14 @@
-import { Filters } from '../../../../Shared/domain/criteria/Filters';
-import { Order } from '../../../../Shared/domain/criteria/Order';
-import { Query } from '../../../../Shared/domain/Query';
-import { QueryHandler } from '../../../../Shared/domain/QueryHandler';
+import { Filters } from '../../../Shared/domain/criteria/Filters';
+import { Order } from '../../../Shared/domain/criteria/Order';
+import { Query } from '../../../Shared/domain/Query';
+import { QueryHandler } from '../../../Shared/domain/QueryHandler';
 import { BackofficeCoursesResponse } from '../BackofficeCoursesResponse';
 import { CoursesByCriteriaSearcher } from './CoursesByCriteriaSearcher';
 import { SearchCoursesByCriteriaQuery } from './SearchCoursesByCriteriaQuery';
 
 export class SearchCoursesByCriteriaQueryHandler
-  implements QueryHandler<SearchCoursesByCriteriaQuery, BackofficeCoursesResponse>
-{
-  constructor(private searcher: CoursesByCriteriaSearcher) {}
+  implements QueryHandler<SearchCoursesByCriteriaQuery, BackofficeCoursesResponse> {
+  constructor(private searcher: CoursesByCriteriaSearcher) { }
 
   subscribedTo(): Query {
     return SearchCoursesByCriteriaQuery;
