+++ src/Contexts/pricingService/domain/MarketClient.ts
@@ -1,0 +1,5 @@
+import { PriceUpdateData } from "../../../Events/PriceUpdateEvent";
+
+export interface MarketClient {
+  fetchMarketPrices(): Promise<PriceUpdateData[]>;
+}
