+++ src/Contexts/pricingService/infrastructure/RabbitMQ/RabbitMQConfigFactory.ts
@@ -1,5 +1,5 @@
-import { ConnectionSettings } from '../../../../Shared/infrastructure/EventBus/RabbitMQ/ConnectionSettings';
-import { ExchangeSetting } from '../../../../Shared/infrastructure/EventBus/RabbitMQ/ExchangeSetting';
+import { ConnectionSettings } from '../../../Shared/infrastructure/EventBus/RabbitMQ/ConnectionSettings';
+import { ExchangeSetting } from '../../../Shared/infrastructure/EventBus/RabbitMQ/ExchangeSetting';
 import config from '../config';
 
 export type RabbitMQConfig = {
