+++ src/Contexts/pricingService/infrastructure/config/dev.json
