+++ src/apps/pricingService/command/ConfigureRabbitMQCommand.ts
@@ -6,11 +6,11 @@ import container from '../dependency-injection';
 
 export class ConfigureRabbitMQCommand {
   static async run() {
-    const connection = container.get<RabbitMqConnection>('Mechoui3.Shared.RabbitMQConnection');
-    const { name: exchange } = container.get<RabbitMQConfig>('Mechoui3.Shared.RabbitMQConfig').exchangeSettings;
+    const connection = container.get<RabbitMqConnection>('PricingService.Shared.RabbitMQConnection');
+    const { name: exchange } = container.get<RabbitMQConfig>('PricingService.Shared.RabbitMQConfig').exchangeSettings;
     await connection.connect();
 
-    const configurer = container.get<RabbitMQConfigurer>('Mechoui3.Shared.RabbitMQConfigurer');
+    const configurer = container.get<RabbitMQConfigurer>('PricingService.Shared.RabbitMQConfigurer');
     const subscribers = DomainEventSubscribers.from(container).items;
 
     await configurer.configure({ exchange, subscribers });
