+++ src/apps/pricingService/dependency-injection/Courses/application.yaml
@@ -1,19 +1,19 @@
 services:
-  Mechoui3.Courses.domain.CourseRepository:
+  PricingService.Courses.domain.CourseRepository:
     class: ../../../../Contexts/Mooc/Courses/infrastructure/persistence/MongoCourseRepository
-    arguments: ['@Mechoui3.Shared.ConnectionManager']
+    arguments: ['@PricingService.Shared.ConnectionManager']
 
-  Mechoui3.Courses.application.CourseCreator:
+  PricingService.Courses.application.CourseCreator:
     class: ../../../../Contexts/Mooc/Courses/application/CourseCreator
-    arguments: ['@Mechoui3.Courses.domain.CourseRepository', '@Mechoui3.Shared.domain.EventBus']
+    arguments: ['@PricingService.Courses.domain.CourseRepository', '@PricingService.Shared.domain.EventBus']
 
-  Mechoui3.courses.Boom:
-    class: ../../../../Contexts/mechoui3/Courses/application/Create/Boom3
+  PricingService.courses.Boom:
+    class: ../../../../Contexts/pricingService/application/Create/Boom3
     tags:
       - { name: 'domainEventSubscriber' }
 
-  Mechoui3.courses.CreateCourseCommandHandler:
+  PricingService.courses.CreateCourseCommandHandler:
     class: ../../../../Contexts/Mooc/Courses/application/Create/CreateCourseCommandHandler
-    arguments: ['@Mechoui3.Courses.application.CourseCreator']
+    arguments: ['@PricingService.Courses.application.CourseCreator']
     tags:
       - { name: 'commandHandler' }
