+++ src/apps/pricingService/dependency-injection/application_test.yaml
@@ -4,4 +4,4 @@ imports:
 services:
   Mooc.EnvironmentArranger:
     class: ../../../../tests/Contexts/Shared/infrastructure/mongo/MongoEnvironmentArranger
-    arguments: ['@Mechoui3.Shared.ConnectionManager']
+    arguments: ['@PricingService.Shared.ConnectionManager']
