+++ src/apps/pricingService/routes/status.route.ts
@@ -4,7 +4,7 @@ import StatusController from '../controllers/StatusGetController';
 
 export const registerStatusRoutes = (): Router => {
   const router = Router();
-  const controller: StatusController = container.get('Apps.mechoui3.controllers.StatusGetController');
+  const controller: StatusController = container.get('Apps.pricingService.controllers.StatusGetController');
   router.get('/status', (req: Request, res: Response) => controller.run(req, res));
 
   return router;
